fix: add option to register with admin or staff role on api/admin/register

+++ src/pages/api/admin/register.ts
@@ -15,6 +15,7 @@ interface AdminRegistrationRequestBody {
   phoneNumber: string;
   email: string;
   password: string;
+  roleId: number;
   agreementAccepted: boolean;
 }
 
@@ -31,6 +32,7 @@ export default async function adminRegister(
     phoneNumber,
     email,
     password,
+    roleId,
     agreementAccepted,
   }: AdminRegistrationRequestBody = req.body;
   // Validate input
@@ -54,13 +56,29 @@ export default async function adminRegister(
   if (!agreementAccepted) {
     res.status(400).json({ error: "You must accept the agreement and policy" });
   }
+  if (roleId < 2 || roleId > 3) {
+    res.status(400).json({ error: "Role id can only be 2 or 3" });
+  }
+
+  let roleTable;
+  let roleColumn;
+  let roleText;
+  if (roleId === 2) {
+    roleTable = "admins";
+    roleColumn = "admin_id";
+    roleText = "admin";
+  } else if (roleId === 3) {
+    roleTable = "staffs";
+    roleColumn = "staff_id";
+    roleText = "staff";
+  }
 
   try {
     //FIXME change to supabase query && find new way to get email
     //check if user already exists in users table
     const { data: checkIfUserExit } = await adminSupabase
-      .from("admins")
-      .select("admin_id")
+      .from(`${roleTable}`)
+      .select(`${roleColumn}`)
       .eq("email", email) //<<<FIXME
       .single();
 
@@ -76,7 +94,7 @@ export default async function adminRegister(
         data: {
           name: name,
           phone: phoneNumber,
-          role: "admin",
+          role: roleText,
         },
       },
     });
@@ -92,34 +110,57 @@ export default async function adminRegister(
     if (!data.user) {
       return res.status(500).json({ error: "Failed to create user" });
     }
-
+    console.log("Check signup I", data);
     // Insert user data into users table
     //FIXME change to supabase query && change user_type to role_id = 2
-    const { data: insertedAdmin, error: insertedAdminError } =
-      await adminSupabase
-        .from("admins")
-        .insert([
-          {
-            admin_id: data.user.id,
-            name,
-            email,
-            phone_number: phoneNumber,
-            address: null,
-            role_id: 2,
-          },
-        ])
-        .select();
+    if (roleId === 2) {
+      const { data: insertedAdmin, error: insertedAdminError } =
+        await adminSupabase
+          .from(`${roleTable}`)
+          .insert([
+            {
+              admin_id: data.user.id,
+              name,
+              email,
+              phone_number: phoneNumber,
+              address: null,
+              role_id: roleId,
+            },
+          ])
+          .select();
 
-    if (insertedAdminError) {
-      await adminSupabase.auth.admin.deleteUser(data.user.id);
-      console.log("Fail to insert admin data.");
-      console.log(insertedAdminError);
-      return res.status(500).json({ error: "Fail to insert admin data." });
+      if (insertedAdminError) {
+        await adminSupabase.auth.admin.deleteUser(data.user.id);
+        console.log("Fail to insert admin data.");
+        console.log(insertedAdminError);
+        return res.status(500).json({ error: "Fail to insert admin data." });
+      }
+    }
+    if (roleId === 3) {
+      const { data: insertedAdmin, error: insertedAdminError } =
+        await adminSupabase
+          .from(`${roleTable}`)
+          .insert([
+            {
+              staff_id: data.user.id,
+              name,
+              email,
+              phone_number: phoneNumber,
+              address: null,
+              role_id: roleId,
+            },
+          ])
+          .select();
+
+      if (insertedAdminError) {
+        await adminSupabase.auth.admin.deleteUser(data.user.id);
+        console.log("Fail to insert admin data.");
+        console.log(insertedAdminError);
+        return res.status(500).json({ error: "Fail to insert admin data." });
+      }
     }
 
-    return res
-      .status(201)
-      .json({ message: "Register successfully", detail: insertedAdmin[0] });
+    return res.status(201).json({ message: "Register successfully" });
   } catch (err) {
     console.log("Unexpected error during registration:", err);
     return res.status(500).json({ error: "An unexpected error occurred" });
